Use docker-compose to provide the env. vars to the dev-server
It looks like the dotenv library either doesn't sufficiently put the
environment variables in place or is too late for when the AWS library
is looking for them. Using the docker-compose.yml file to provide the
environment variable (via that same development/development.env file)
seems to have satisfied the AWS library.